Move classgroup.py in fast_vdf to util.
